requirements.txt
Django==4.0.6
djangorestframework==3.13.1
mysqlclient==2.1.1
pymysql==1.0.2
django-redis==5.2.0
Pillow==9.2.0
cryptography==37.0.4
requests==2.28.1
celery==5.2.7
eventlet==0.33.2